refac: 로그인 버튼 클릭 시 최적화
