+++ src/components/molecules/Div/LoginModalFooter.tsx
@@ -1,16 +1,39 @@
 import DefaultButton from "@components/atoms/Button/DefaultButton";
 import LinkText from "@components/atoms/Text/LinkText";
 import Text from "@components/atoms/Text/Text";
+import useLogin from "@hooks/query/useLogin";
+import validateEmail from "@utils/validateEmail";
+import { Dispatch, Ref, SetStateAction, useCallback } from "react";
 import { useNavigate } from "react-router-dom";
 
 interface LoginModalFooterProps {
-  onLogin: () => void;
+  refs: any;
+  setErrorMessage: Dispatch<SetStateAction<string>>;
 }
 
-const LoginModalFooter = ({ onLogin }: LoginModalFooterProps) => {
+const LoginModalFooter = ({ refs, setErrorMessage }: LoginModalFooterProps) => {
   const navigate = useNavigate();
+  const { mutate: login } = useLogin(setErrorMessage);
+  const { emailRef, passwordRef } = refs;
 
-  const goRegisterPage = () => navigate("/register");
+  const goRegisterPage = useCallback(() => navigate("/register"), []);
+
+  const onLogin = useCallback(() => {
+    if (!emailRef.current.value || !passwordRef.current.value) {
+      return setErrorMessage("모든 값을 입력해주세요.");
+    }
+    if (!validateEmail(emailRef.current.value)) {
+      return setErrorMessage("유효하지 않은 아이디입니다.");
+    }
+    if (passwordRef.current.value.length < 8) {
+      return setErrorMessage("비밀번호는 8자리 이상이어야 합니다.");
+    }
+    setErrorMessage("");
+    login({
+      email: emailRef.current.value,
+      password: passwordRef.current.value,
+    });
+  }, []);
 
   return (
     <>
